generateUniqId moved to utilities
